Rename tcpfwd filename to forward
Now handles stream forwarding as well
